+++ crates/ironcalc-base/src/expressions/parser/static_analysis.rs
@@ -896,6 +896,7 @@ fn get_function_args_signature(kind: &Function, arg_count: usize) -> Vec<Signatu
         Function::Bahttext => args_signature_scalars(arg_count, 1, 0),
         Function::Arraytotext => args_signature_scalars(arg_count, 1, 1),
         Function::Transpose => vec![Signature::Vector],
+        Function::Sequence => args_signature_scalars(arg_count, 1, 3),
         Function::Aggregate => {
             let mut result = vec![Signature::Vector; arg_count];
             if arg_count >= 1 {
@@ -1485,6 +1486,7 @@ fn static_analysis_on_function(kind: &Function, args: &[Node]) -> StaticResult {
         Function::Bahttext => not_implemented(args),
         Function::Arraytotext => not_implemented(args),
         Function::Transpose => not_implemented(args),
+        Function::Sequence => not_implemented(args),
         Function::Aggregate => not_implemented(args),
         Function::Rand => not_implemented(args),
         Function::Randbetween => scalar_arguments(args),